Update code style
