+++ src/main/java/lax/command/ExitCommand.java
@@ -1,16 +1,16 @@
-package Lax.Command;
+package lax.command;
 
-import Lax.Storage;
-import Lax.Task.TaskList;
-import Lax.Ui.Ui;
+import lax.Storage;
+import lax.task.TaskList;
+import lax.ui.Ui;
 
 /**
- * Represents a list command.
+ * Represents an exit command.
  */
-public class ListCommand extends Command {
+public class ExitCommand extends Command {
     /**
      * {@inheritDoc}
-     * It displays the full tasklist to the user.
+     * It sets the exit status of the command to <code>true</code> and closes the chatbot.
      *
      * @param taskList The tasklist to modify.
      * @param ui       The ui for displaying messages to the user.
@@ -18,6 +18,7 @@ public class ListCommand extends Command {
      */
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
-        ui.showList(taskList);
+        ui.showExit();
+        super.setExit(true);
     }
 }
