+++ src/main/java/lax/command/LabelCommand.java
@@ -1,10 +1,10 @@
-package Lax.Command;
+package lax.command;
 
-import Lax.Exception.InvalidCommandException;
-import Lax.Storage;
-import Lax.Task.Task;
-import Lax.Task.TaskList;
-import Lax.Ui.Ui;
+import lax.Storage;
+import lax.exception.InvalidCommandException;
+import lax.task.Task;
+import lax.task.TaskList;
+import lax.ui.Ui;
 
 /**
  * Represents a label command with a <code>String</code> taskNumber and <code>boolean</code> mark.
